+++ api-rework/invasives/api/viewsets/__init__.py
@@ -5,4 +5,3 @@ from .maps import MapGenerationRecordViewset, MapGenerationRequestViewSet
 from .migration import MigrationStatusViewSet
 from .recordset_rows import RecordsetRowsViewSet
 from .vector_tiles import VectorTileViewset
-from .exports import ExportViewset
